actualizacion ejercicio 6 de while

+++ 5-InstruccionWhile/jsIteracionesWhile-06.js
@@ -13,7 +13,7 @@ function mostrar()
 	while (condicion < 5) {
 
 		numeroIngresado = prompt("Ingrese un numero");
-		numeroIngresado = parseFloat(numeroIngresado);
+		numeroIngresado = parseInt(numeroIngresado);
 
 		acumulador = acumulador + numeroIngresado;
